backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
databases[postgresql]==0.7.1
SQLAlchemy==2.1.2
asyncpg==0.27.1
python-dotenv==1.0.0
pydantic==2.5.1
fastapi-jwt-auth==0.7.0
passlib[bcrypt]==1.7.5
python-multipart==0.0.6
